requirements.txt
numpy>=1.26
pillow>=10.2
opencv-python>=4.9
torch>=2.2
torchvision>=0.17
streamlit>=1.32
scikit-learn>=1.4

